Replace InputNotes<InputNote> with vec<Note>

+++ crates/miden-tx/src/errors/mod.rs
@@ -28,7 +28,7 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum NoteCheckerError {
     #[error("invalid input note count {0} is out of range)")]
-    InputNoteCountOutOfRange(u16),
+    InputNoteCountOutOfRange(usize),
     #[error("transaction preparation failed: {0}")]
     TransactionPreparation(#[source] TransactionExecutorError),
     #[error("transaction execution prologue failed: {0}")]
